Sistemazioni nelle classi in modo da poter avviare il gioco.
Inserite nella cartella immagini l'immagine del tabellone (provvisioria)
e un'altra cartella contenente le faccie dei dadi.
Sviluppata interfaccia grafica della FinestraDiGioco suddividendo in base a righe e colonne il tabellone.
Implementato sorteggio ed estrazione dei dadi sia a codice che in interfaccia.

+++ ProgettoMonopoly/ProgettoMonopoly/Classi/Contratti/ContrattoStazione.cs
@@ -7,7 +7,7 @@ namespace ProgettoMonopoly
 {
     public class ContrattoStazione : Contratto
     {
-        public ContrattoStazione(string nomeContratto, float valoreContratto, List<float> rendita) : base(nomeContratto, valoreContratto, rendita)
+        public ContrattoStazione(string nomeContratto, float valoreContratto, List<int> rendita) : base(nomeContratto, valoreContratto, rendita)
         {
 
         }
